requirements.txt
langchain
langchain_community
openai
chainlit
python-dotenv
